+++ models/userModel.js
@@ -1,0 +1,16 @@
+exports.user = {
+    "id":null,
+    "fullName":null,
+    "phoneNumber":null,    
+    "email":null,
+    "password":null,
+    "dateCreated":null,
+    "lastUpdate":null,
+    "donations":null,
+    "requests":null,
+    "lastSeen":null,
+    "accountType":null,
+    "verified":null,
+    "token":null,
+    "deviceDetail":null
+}