+++ models/utilitiesModel.js
@@ -5,7 +5,22 @@ exports.responseObj = {
     "data":null
 }
 
-exports.headers = {'Content-Type': 'application/json'}
+exports.headers = {
+    'Content-Type': 'application/json',
+    'Access-Control-Allow-Origin':'*',  
+    'Access-Control-Allow-Headers': 'Origin, X-Requested-With, Content-Type, Accept, Authorization',
+    'Access-Control-Allow-Methods': 'POST,OPTIONS',
+    'Access-Control-Allow-Credentials': true
+}
+
+exports.firebaseConfig = {
+    apiKey: "AIzaSyA1yiQ3hn5sJjva5DPLzCNcD3JkGEyI9Dk",
+    authDomain: "rich-out.firebaseapp.com",
+    databaseURL: "https://rich-out.firebaseio.com",
+    projectId: "rich-out",
+    storageBucket: "rich-out.appspot.com",
+    messagingSenderId: "550404557767"
+};
 
 exports.resCodes = {
     'function_not_found': {
@@ -20,9 +35,17 @@ exports.resCodes = {
         "code":400,     
         "message":"An error occured while performing request"  
     },
+    'invalid_data':{
+        "code":400,     
+        "message":"Invalid data format"  
+    },
     'request_succesful':{
         "code":200,
         "message":"Request Successful"
+    },
+    'route_not_found':{
+        "code":500,
+        "message":"Cannot find route"
     },
     '500': {
         "code":500,
@@ -34,8 +57,22 @@ exports.resCodes = {
     },
     '401': {
         "code":401,
-        "message":""
+        "message":"API Authentication Failed"
+    },
+    '204': {
+        "code":204,
+        "message":"Handle Options CORS Request"
     }
 }
 
+exports.appConfig={
+    appState:'test',
+    liveHostName:'0.0.0.0',
+    livePort:3000,
+    testHostName:'127.0.0.1',
+    testPort:3000,
+    apiUser:"am9objpzbWl0aA==",
+    apiPass:"JiZAQEFBMTE6NjcmOCMh"
+
+}
 //